Refactor setup

+++ setup-main.sh
@@ -1,40 +1,35 @@
 #!/bin/sh
 
 ###############################################################################
+# Helpers
+
+link_file() {
+    local src=$1
+    local target=$2
+
+    echo "Set: '${src}' -> '${target}'"
+    
+    if [ -e "$target" ]; then
+        rm -rf "$target"
+    fi
+
+    ln -nfs "$src" "$target"
+}
+
+###############################################################################
 # General configs
 
 echo
 echo "➡️ Setting up general configs..."
 
-install_dir() {
-    for src in "$1"/*; do
-        install_link "$src"
+install_dot_configs() {
+    for src in $(pwd)/dot/*; do
+        local target="$HOME/.$(basename "$src")"
+        link_file "$src" "$target"
     done
 }
 
-install_link() {
-    if [ -e "$1" ]; then
-        full_source=$(pwd)/$1
-        file=$(basename "$1")
-        target="$HOME/.$file"
-
-        echo "Set: '${full_source}' -> '${target}'"
-        
-        # Backup and copy
-        if [ -e "$target" ]; then
-            backup_target="${target}.backup"
-            echo "[Overwriting] '${target}'. Leaving original at '${backup_target}'"
-            mv "$target" "$backup_target"
-        fi
-
-        ln -nfs "$full_source" "$target"
-    fi
-}
-
-install_dir "git"
-install_dir "ruby"
-install_dir "vimify"
-install_link "vim"
+install_dot_configs
 
 ###############################################################################
 # Package manager
@@ -43,15 +38,14 @@ echo
 echo "➡️ Setting up brew..."
 
 install_brew() {
-    # Install command-line tools using Homebrew.
+    if test ! $(which brew); then
+        /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
 
-    brew || /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
-
-    # Make sure we’re using the latest Homebrew.
-    brew update
-
-    # Upgrade any already-installed formulae.
-    brew upgrade
+        brew update
+        brew upgrade
+    else
+        echo "Brew is already installed."
+    fi
 }
 
 install_brew
@@ -62,37 +56,32 @@ install_brew
 echo
 echo "➡️ Setting up Fish shell..."
 
-brew install fish
+install_fish() {
+    if [ ! -f "/opt/homebrew/bin/fish" ]; then
+        brew install fish
+    fi
 
-install_fish_config() {
-    for src in $(pwd)/fish/*; do
-        file=$(basename "${src}")
-        target="$HOME/.config/fish/$file"
+    if [ ! -f "/opt/homebrew/bin/starship" ]; then
+        brew install starship
+    fi
 
-        echo "Set: '${src}' -> '${target}'"
+    # Fish config
 
-        if [ -e "$target" ]; then
-            rm -rf "$target"
-        fi
-
-        ln -nfs "$src" "$target"
+    for src in $(pwd)/config/fish/*; do
+        local target="$HOME/.config/fish/$(basename "${src}")"
+        link_file "$src" "$target"
     done
-}
 
-install_fish_config
+    # Set fish as default shell
 
-install_fish_shell() {
-    # Check if Fish is installed at /opt/homebrew/bin/fish
     if [ -f "/opt/homebrew/bin/fish" ]; then
         echo "Fish shell found at /opt/homebrew/bin/fish"
 
-        # Check if Fish is already in the list of standard shells
         if ! grep -q "/opt/homebrew/bin/fish" /etc/shells; then
             echo "Adding Fish to the list of standard shells."
             echo "/opt/homebrew/bin/fish" | sudo tee -a /etc/shells > /dev/null
         fi
 
-        # Check if the current shell is already Fish
         if [ "$SHELL" = "/opt/homebrew/bin/fish" ]; then
             echo "Fish is already the default shell."
         else
@@ -102,20 +91,23 @@ install_fish_shell() {
     else
         echo "Fish shell not found at /opt/homebrew/bin/fish. Please install Fish and try again."
     fi
+
+    # Fish plugins
+
+    fish -c 'curl -sL https://raw.githubusercontent.com/jorgebucaran/fisher/main/functions/fisher.fish | source && fisher install jorgebucaran/fisher'
+    fish -c 'fisher install jorgebucaran/nvm.fish'
+    fish -c 'fisher install oh-my-fish/plugin-foreign-env'
+
+    # Misc
+
+    link_file "$(pwd)/config/starship.toml" "$HOME/.config/starship.toml"
+    link_file "$(pwd)/config/gitui" "$HOME/.config/gitui"
+    
+    # Disable Last Login Message
+    touch $HOME/.hushlogin 
 }
 
-install_fish_shell
-
-install_fish_plugins() {
-    /opt/homebrew/bin/fish
-
-    curl -sL https://raw.githubusercontent.com/jorgebucaran/fisher/main/functions/fisher.fish | source && fisher install jorgebucaran/fisher
-    fisher install jorgebucaran/nvm.fish
-
-    exit
-}
-
-install_fish_plugins
+install_fish
 
 echo
 echo "✅ dotfiles has been installed. Please restart your terminal."
